requirements.txt
asyncpg
aiogram==2.15
python-dotenv
asyncio